full graph functionality

+++ graph/graph_getting_array_value_test.go
@@ -14,14 +14,14 @@ func TestGetArrayValue_1(t *testing.T) {
 	diagram := graph.NewGraph(nil)
 	allVars := map[string]any{
 		"x":   3.0,
-		"tab": []float32{3.0, 4.0, 1.0, 2.0, 3.0},
+		"tab": []float64{3.0, 4.0, 1.0, 2.0, 3.0},
 	}
 	diagram.SetAllVars(allVars)
 	kvp, err := diagram.GetKvpByKeys(&[]string{"x", "tab[3]"})
 	if err != nil {
 		t.Errorf("Error: '%s'", err.Error())
 	}
-	targetKvp := map[string]float32{
+	targetKvp := map[string]float64{
 		"x":      3.0,
 		"tab[3]": 2.0,
 	}
@@ -34,7 +34,7 @@ func TestGetArrayValue_2(t *testing.T) {
 	diagram := graph.NewGraph(nil)
 	allVars := map[string]any{
 		"x":   3.0,
-		"tab": []float32{3.0, 4.0, 1.0, 2.0, 3.0},
+		"tab": []float64{3.0, 4.0, 1.0, 2.0, 3.0},
 		"y":   0.2,
 	}
 	diagram.SetAllVars(allVars)
@@ -42,7 +42,7 @@ func TestGetArrayValue_2(t *testing.T) {
 	if err != nil {
 		t.Errorf("Error: '%s'", err.Error())
 	}
-	targetKvp := map[string]float32{
+	targetKvp := map[string]float64{
 		"tab[2 * x - 2]": 3.0,
 	}
 	if !reflect.DeepEqual(targetKvp, kvp) {
@@ -54,7 +54,7 @@ func TestGetArrayValue_3(t *testing.T) {
 	diagram := graph.NewGraph(nil)
 	allVars := map[string]any{
 		"x":   3.0,
-		"tab": []float32{3.0, 4.0, 1.0, 2.0, 3.0},
+		"tab": []float64{3.0, 4.0, 1.0, 2.0, 3.0},
 		"y":   0.2,
 	}
 	diagram.SetAllVars(allVars)
@@ -71,7 +71,7 @@ func TestGetArrayValue_4(t *testing.T) {
 	diagram := graph.NewGraph(nil)
 	allVars := map[string]any{
 		"x":   3.0,
-		"tab": []float32{3.0, 4.0, 1.0, 2.0, 3.0},
+		"tab": []float64{3.0, 4.0, 1.0, 2.0, 3.0},
 		"y":   0.2,
 	}
 	diagram.SetAllVars(allVars)
